+++ pyoxidizer/src/project_layout.rs
@@ -20,6 +20,15 @@ lazy_static! {
         let mut handlebars = Handlebars::new();
 
         handlebars
+            .register_template_string(
+                "application-manifest.rc",
+                include_str!("templates/application-manifest.rc"),
+            )
+            .unwrap();
+        handlebars
+            .register_template_string("exe.manifest", include_str!("templates/exe.manifest"))
+            .unwrap();
+        handlebars
             .register_template_string("new-build.rs", include_str!("templates/new-build.rs"))
             .unwrap();
         handlebars
@@ -138,8 +147,9 @@ pub fn write_new_cargo_config(project_path: &Path) -> Result<()> {
     Ok(())
 }
 
-pub fn write_new_build_rs(path: &Path) -> Result<()> {
-    let data: BTreeMap<String, String> = BTreeMap::new();
+pub fn write_new_build_rs(path: &Path, program_name: &str) -> Result<()> {
+    let mut data = TemplateData::new();
+    data.program_name = Some(program_name.to_string());
     let t = HANDLEBARS.render("new-build.rs", &data)?;
 
     println!("writing {}", path.display());
@@ -186,6 +196,33 @@ pub fn write_new_pyoxidizer_config_file(
     println!("writing {}", path.to_str().unwrap());
     let mut fh = std::fs::File::create(path)?;
     fh.write_all(t.as_bytes())?;
+
+    Ok(())
+}
+
+/// Write an application manifest and corresponding resource file.
+///
+/// This is used on Windows to allow the built executable to use long paths.
+///
+/// Windows 10 version 1607 and above enable long paths by default. So we
+/// might be able to remove this someday. It isn't clear if you get long
+/// paths support if using that version of the Windows SDK or if you have
+/// to be running on a modern Windows version as well.
+pub fn write_application_manifest(project_dir: &Path, program_name: &str) -> Result<()> {
+    let mut data = TemplateData::new();
+    data.program_name = Some(program_name.to_string());
+
+    let manifest_path = project_dir.join(format!("{}.exe.manifest", program_name));
+    let manifest_data = HANDLEBARS.render("exe.manifest", &data)?;
+    println!("writing {}", manifest_path.display());
+    let mut fh = std::fs::File::create(&manifest_path)?;
+    fh.write_all(manifest_data.as_bytes())?;
+
+    let rc_path = project_dir.join(format!("{}-manifest.rc", program_name));
+    let rc_data = HANDLEBARS.render("application-manifest.rc", &data)?;
+    println!("writing {}", rc_path.display());
+    let mut fh = std::fs::File::create(&rc_path)?;
+    fh.write_all(rc_data.as_bytes())?;
 
     Ok(())
 }
@@ -282,6 +319,10 @@ pub fn update_new_cargo_toml(path: &Path, pyembed_location: &PyembedLocation) ->
     ));
 
     content.push_str("\n");
+    content.push_str("[build-dependencies]\n");
+    content.push_str("embed-resource = \"1.3\"\n");
+
+    content.push_str("\n");
     content.push_str("[features]\n");
     content.push_str("default = [\"build-mode-pyoxidizer-exe\"]\n");
     content.push_str("jemalloc = [\"jemallocator-global\", \"pyembed/jemalloc\"]\n");
@@ -323,9 +364,10 @@ pub fn initialize_project(
     add_pyoxidizer(&path, true)?;
     update_new_cargo_toml(&path.join("Cargo.toml"), pyembed_location)?;
     write_new_cargo_config(&path)?;
-    write_new_build_rs(&path.join("build.rs"))?;
+    write_new_build_rs(&path.join("build.rs"), name)?;
     write_new_main_rs(&path.join("src").join("main.rs"))?;
     write_new_pyoxidizer_config_file(&path, &name, code, pip_install)?;
+    write_application_manifest(&path, &name)?;
 
     Ok(())
 }